% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntc.R
\name{ntc}
\alias{ntc}
\title{Name that color}
\usage{
ntc(colors)
}
\arguments{
\item{the}{color to be named}
}
\value{
a list with the code, the name, the RGB and HSL values of the closest
matching color.
}
\description{
Name that color
}

