% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb2hsl.R
\name{rgb2hsl}
\alias{rgb2hsl}
\title{RGB to HSL}
\usage{
rgb2hsl(rgb)
}
\arguments{
\item{rgb}{a numeric vector of length 3 specifying red, green and blue in
that order}
}
\value{
a numeric vector of length 3 named "h", "s", "l"
}
\description{
RGB to HSL
}

